Move start-lighthouse to separate file
